backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-ai
pydantic-settings
sse-starlette
python-dotenv
httpx
tavily-python
